Make it a little more difficult to accidentally delete entries.

+++ assets/tags/entry-view.tag
@@ -10,17 +10,27 @@
   <pre class="entry-text">{opts.entry.text}</pre>
   <div if="{opts.entry.isOwner}" class="entry-actions">
     <form method="post" action="/entry/{opts.entry.id}?_method=DELETE" onsubmit="{del}" class="form-submit-only">
-      <input type="submit" value="Delete" class="pure-button button-warning action entry-action--left"/>
+      <input
+        type="submit"
+        value="{opts.delText}"
+        class="pure-button {'button-warning': !opts.delClickCount, 'button-error': opts.delClickCount} action entry-action--left"/>
     </form>
     <a href="/entry/{opts.entry.id}/edit" onclick="{edit}" class="pure-button button-margin-left action entry-action--right">Edit</a>
   </div>
 
   <script>
     var self = this;
+    opts.delText = 'Delete';
+    opts.delClickCount = 0;
     self.edit = function(e){
       opts.page('/entry/' + opts.entry.id + '/edit', {data: opts.entry});
     }
     self.del = function(e){
+      ++opts.delClickCount;
+      if(opts.delClickCount === 1){
+        opts.delText = 'Confirm';
+        return self.update();
+      }
       opts.entryService.deleteEntry(opts.entry.id).then(function (response){
         window.deleteEntriesKeys();
         opts.page.replace('/entries');
